Created BaseEntity.java

+++ src/main/java/com/dalfatih/hrms/dataAccess/abstracts/JobCategoryRepository.java
@@ -4,5 +4,5 @@ import com.dalfatih.hrms.entities.concretes.JobCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface JobCategoryRepository extends JpaRepository<JobCategory, Integer> {
+public interface JobCategoryRepository extends JpaRepository<JobCategory, Long> {
 }
